pyproject.toml
[tool.poetry]
name = "lnspec-codegen"
version = "0.1.0"
description = ""
authors = ["Vincenzo Palazzo <vincenzopalazzodev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyln-proto = "^0.10.1"
black = "^21.8b0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
